use `BitmapObjectPool` for `SquashFS` readers
Finish replacing ConcurrentQueue with BitmapObjectPool by rewriting SquashFS support.
The pool recently got support to pass arbitrary parameters to acquire_scoped(), use it to pass references to the archives (archives are stored in a vector, so we can't promise their
addresses won't change).

Refactor my own code which I wrote half a year ago to be more C++
my experience has evolved a lot lol. Use std::unique_ptr<> and wrap_unique() for everything we can to delegate resource management to the constructors and destructors.

Misc changes:
* declare the default deleter for CStreamReader the same way as for  IReader to be able to safely use wrap_unique();
* wrap squashfs-tools-ng's C headers into `sqfs` namespace for
  better isolation;

xrCore: optimize and make SquashFS readers fully thread-safe:
libsquashfs from squashfs-tools-ng doesn't care about
multi-threading and doesn't provide any guarantees. That's why I
initially made a "reader" class which always has its own file ref,
directory, and file readers.
What I missed is that sqfs_compressor_t should not be shared as well.

Make readers have their own compressor to avoid the refcount mess.
Move the compressor config struct to the pool to initialize
compressors faster. compressor_config_t, as its name says, contains
only configuration options, no pointers or refcounts, so it can be
shared the same way as we share sqfs_super_t.
This also simplifies the entire logic -- no more global sqfs_file_t,
every operation must be done using a reader. The initial file used
to read the superblock gets destroyed right after the operation.

+++ ogsr_engine/xrCore/fs_internal.h
@@ -45,7 +45,7 @@ public:
         }
     }
     // kernel
-    virtual void w(const void* _ptr, u32 count)
+    virtual void w(const void* _ptr, size_t count)
     {
         if ((0 != hf) && (0 != count))
         {
@@ -53,21 +53,21 @@ public:
             R_ASSERT3(W == count, "Can't write mem block to file. Disk maybe full.", _sys_errlist[errno]);
         }
     };
-    virtual void seek(u32 pos)
+    virtual void seek(size_t pos)
     {
         if (0 != hf)
             fseek(hf, pos, SEEK_SET);
     };
-    virtual u32 tell() { return (0 != hf) ? ftell(hf) : 0; };
+    virtual size_t tell() { return (0 != hf) ? ftell(hf) : 0; };
     virtual bool valid() { return (0 != hf); }
-    virtual int flush() { return fflush(hf); }
+    virtual size_t flush() { return fflush(hf); }
 };
 
 // It automatically frees memory after destruction
 class CTempReader : public IReader
 {
 public:
-    CTempReader(void* _data, int _size, int _iterpos) : IReader(_data, _size, _iterpos) {}
+    CTempReader(void* _data, size_t _size, size_t _iterpos) : IReader(_data, _size, _iterpos) {}
     virtual ~CTempReader();
 };
 
@@ -76,7 +76,7 @@ class CPackReader : public IReader
     void* base_address;
 
 public:
-    CPackReader(void* _base, void* _data, int _size) : IReader(_data, _size) { base_address = _base; }
+    CPackReader(void* _base, void* _data, size_t _size) : IReader(_data, _size) { base_address = _base; }
     virtual ~CPackReader();
 };
